chore(wrapper): extract readVersion to common util (#80)
* chore(wrapper): extract readVersion to common util

Fixes #79

Extracted the duplicated readVersion logic from CodexProvider and GeminiProvider into a shared utility function.

* fix(review): use promisified execFile for readVersion
